// src/contentful.d.ts
// THIS FILE IS AUTOMATICALLY GENERATED. DO NOT MODIFY IT.

import { Asset, Entry } from "contentful";

export interface ICallToActionButtonFields {
  /** Internal Name */
  internalName?: string | undefined;

  /** Call To Action Text */
  callToActionText: string;
}

export interface ICallToActionButton extends Entry<ICallToActionButtonFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "callToActionButton";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IImageWrapperFields {
  /** Internal Name */
  internalName?: string | undefined;

  /** Alt Text */
  altText: string;

  /** Image */
  image: Asset;
}

export interface IImageWrapper extends Entry<IImageWrapperFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "imageWrapper";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface ILandingPageFields {
  /** Internal Name */
  internalName?: string | undefined;

  /** Title */
  title: string;

  /** Slug */
  slug: string;

  /** Sections */
  sections?: (IWelcomeSection | IServicesSection)[] | undefined;
}

export interface ILandingPage extends Entry<ILandingPageFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "landingPage";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IServicesCardFields {
  /** Internal Name */
  internalName?: string | undefined;

  /** Title */
  title: string;

  /** Image */
  image: IImageWrapper;
}

export interface IServicesCard extends Entry<IServicesCardFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "servicesCard";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IServicesSectionFields {
  /** InternalName */
  internalName?: string | undefined;

  /** Title */
  title: string;

  /** Image */
  image: IImageWrapper;

  /** Cards */
  cards: IServicesCard[];

  /** Call To Action */
  callToAction?: ICallToActionButton | undefined;
}

export interface IServicesSection extends Entry<IServicesSectionFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "servicesSection";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IWelcomeSectionFields {
  /** Internal Name */
  internalName?: string | undefined;

  /** Title */
  title: string;

  /** Introduction */
  introduction: string;

  /** Video Embed Url */
  videoEmbedUrl: string;

  /** Call To Action */
  callToAction?: ICallToActionButton | undefined;
}

export interface IWelcomeSection extends Entry<IWelcomeSectionFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "welcomeSection";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export type CONTENT_TYPE =
  | "callToActionButton"
  | "imageWrapper"
  | "landingPage"
  | "servicesCard"
  | "servicesSection"
  | "welcomeSection";

export type IEntry =
  | ICallToActionButton
  | IImageWrapper
  | ILandingPage
  | IServicesCard
  | IServicesSection
  | IWelcomeSection;

export type LOCALE_CODE = "en-US";

export type CONTENTFUL_DEFAULT_LOCALE_CODE = "en-US";